+++ includes/class-plugin.php
@@ -958,6 +958,10 @@ class Plugin {
             return;
         }
 
+        if ($this->incompatible_content_type()) {
+            return;
+        }
+
         $message = sprintf(
             'Performance optimized by Redis Object Cache. Learn more: %s',
             'https://wprediscache.com'
@@ -985,6 +989,51 @@ class Plugin {
             $message, // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
             esc_html( $debug )
         );
+    }
+
+    /**
+     * Whether incompatible content type headers were sent.
+     *
+     * @return bool
+     */
+    protected function incompatible_content_type()
+    {
+        $jsonContentType = static function ($headers) {
+            foreach ($headers as $header => $value) {
+                if (stripos((string) $header, 'content-type') === false) {
+                    continue;
+                }
+
+                if (stripos((string) $value, '/json') === false) {
+                    continue;
+                }
+
+                return true;
+            }
+
+            return false;
+        };
+
+        if (function_exists('headers_list')) {
+            $headers = [];
+
+            foreach (headers_list() as $header) {
+                [$name, $value] = explode(':', $header);
+                $headers[$name] = $value;
+            }
+
+            if ($jsonContentType($headers)) {
+                return true;
+            }
+        }
+
+        if (function_exists('apache_response_headers')) {
+            if ($headers = apache_response_headers()) {
+                return $jsonContentType($headers);
+            }
+        }
+
+        return false;
     }
 
     /**